refactor(git): move print_diff_help to diff module; promote PARALLEL_THRESHOLD
- Move print_diff_help() from git/mod.rs into diff/mod.rs as a private
  function, matching the pattern of all other subcommand help functions
  living in their own module (fixes cohesion issue, gitmod:94).
- Update --mode help text to document that -m has no short flag and
  conflicts with git's own -m flag (gitmod:96).
- Add visibility comment on private submodule declarations in git/mod.rs
  to document the intentional design choice (gitmod:13).
- Promote PARALLEL_THRESHOLD from inline function-body const to module
  level alongside MAX_AST_FILE_SIZE and MAX_AST_FILE_COUNT, and add a
  test documenting its value (diffmod:222).
- Remove intermediate .collect::<Vec<_>>() before .par_iter() by calling
  .par_iter().enumerate() directly on the slice, eliminating an
  unnecessary heap allocation (diffmod:203).
